conformity calc component separated
